requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
python-dotenv==1.1.1
gunicorn==22.0.0
